Add GigaViewer to multisrc (#7125)
* Add GigaViewer to multisrc.

* Add MagComi.
