--- src/ja/shonenjumpplus/build.gradle
@@ -1,12 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-
-ext {
-    extName = 'Shonen Jump+'
-    pkgNameSuffix = 'ja.shonenjumpplus'
-    extClass = '.ShonenJumpPlus'
-    extVersionCode = 2
-    libVersion = '1.2'
-}
-
-apply from: "$rootDir/common.gradle"
